refactor: 修改注释,重构代码
