+++ Program.cs
@@ -4,14 +4,14 @@ using Microsoft.AspNetCore.Http.Connections;
 using SimpleWebChatApplication.Hubs;
 using SimpleWebChatApplication.Services;
 
-if (args.Length == 2 && args[0] == "install") { // install <Password>
+if (args is ["install", _]) { // install <Password>
 	Console.WriteLine("请保证你的密码强度足够，否则可能会被破解！");
 	var password = args[1].Trim();
-	if (!ICheckingTools.IsPasswordComplicated(password)) {
+	if (!IGeneralTools.IsPasswordComplicated(password)) {
 		Console.WriteLine("密码强度不足！");
 		return;
 	}
-	var hash = ICheckingTools.HashPassword(password, out var salt);
+	var hash = IGeneralTools.HashPassword(password, out var salt);
 	Console.WriteLine($"密码为：{password}");
 	Console.WriteLine($"密码的哈希值为：{Convert.ToBase64String(hash)}");
 	Console.WriteLine($"密码的盐值为：{Convert.ToBase64String(salt)}");
